Refactor to match Pygame (and make smaller, more readable source files). Add get_ticks() and perf_counter.
