+++ sdlg/time.py
@@ -1,16 +1,35 @@
-import time as t
+from __future__ import division
+
+# import time as t
 import sdl3
+
+from sdlg import (
+    _perf_to_ms_divisor,
+    _start_perf_count,
+)
+
+
+def get_ticks():
+    if _start_perf_count is None:
+        return 0  # 0 until init as per Pygame
+    return (sdl3.SDL_GetPerformanceCounter() - _start_perf_count) / _perf_to_ms_divisor
+
+
+def get_seconds():
+    if _start_perf_count is None:
+        return 0  # 0 until init as per Pygame
+    return (sdl3.SDL_GetPerformanceCounter() - _start_perf_count) / sdl3.SDL_GetPerformanceFrequency()
 
 
 class Clock:
     def __init__(self):
         self.time = sdl3.SDL_GetPerformanceCounter()
         self.last_time = sdl3.SDL_GetPerformanceCounter()
+        self.freq = sdl3.SDL_GetPerformanceFrequency()
         self.delta = 0
         self.maxFPS = -1
         self.elapsed = 0
         self.draw = 1
-        self.freq = sdl3.SDL_GetPerformanceFrequency()
 
     def tick(self, cap = -1):
         if cap > 0:
@@ -24,7 +43,7 @@ class Clock:
                 self.elapsed -= (1 / self.maxFPS)
                 self.draw = 0
 
-        #self.last_time = self.time
+        # self.last_time = self.time
 
     def get_fps(self):
         self.last_time = self.time
